Добавляет дз.5
